Rename UploadedSong component to SongUploaded
